requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
Flask-Login>=0.6
pyotp>=2.9
argon2-cffi>=23.1
cryptography>=41.0
email-validator>=2.1
python-dotenv>=1.0
click>=8.1
